Cap # of companies a user can watch
