+++ web/app/util/utils.js
@@ -24,6 +24,31 @@ class Utils {
   isFunc(func) {
     return typeof func == 'function';
   }
+
+  /**
+   * Searches the stocks for a company name matching (case-insensitive) the company.
+   * Returns undefined if not found
+   * @param {stock[]} stocks - The list of stocks
+   * @param {string} company - The name of the company to look for
+   * @returns {stock|undefined}
+   */
+  findStockDatum(stocks, company) {
+
+    if (!company) {
+      return undefined;
+    }
+
+    for (var i=0; i<stocks.length; i++) {
+      var stock = stocks[i];
+      var name = (stock.company && stock.company.toLowerCase()) || "";
+      if (name === company.toLowerCase()) {
+        return stock;
+      }
+    }
+
+    return undefined;
+  }
+
 }
 
 module.exports = new Utils()
